tests: clean up benchmark code
