+++ decoder/decoder_test.go
@@ -31,7 +31,6 @@ import (
 	"github.com/muktihari/fit/kit/hash/crc16"
 	"github.com/muktihari/fit/profile"
 	"github.com/muktihari/fit/profile/basetype"
-	"github.com/muktihari/fit/profile/filedef"
 	"github.com/muktihari/fit/profile/mesgdef"
 	"github.com/muktihari/fit/profile/typedef"
 	"github.com/muktihari/fit/profile/untyped/fieldnum"
@@ -2661,144 +2660,4 @@ func BenchmarkDecodeMessageData(b *testing.B) {
 		}
 		cur = 0 // reset reader
 	}
-}
-
-func BenchmarkDecode(b *testing.B) {
-	b.StopTimer()
-	// This is not a typical FIT in term of file size (2.3M) and the messages size.
-	// But since it's big, it's should be good to benchmark.
-	f, err := os.Open("../testdata/big_activity.fit")
-	// f, err := os.Open("../testdata/from_official_sdk/activity_lowbattery.fit")
-	if err != nil {
-		panic(err)
-	}
-	defer f.Close()
-
-	all, err := io.ReadAll(f)
-	if err != nil {
-		panic(err)
-	}
-
-	cur := 0
-	r := fnReader(func(b []byte) (n int, err error) {
-		if cur == len(all) {
-			return 0, io.EOF
-		}
-		n = copy(b, all[cur:])
-		cur += n
-		return
-	})
-
-	dec := New(r)
-	b.StartTimer()
-
-	for i := 0; i < b.N; i++ {
-		_, err = dec.Decode()
-		if err != nil {
-			b.Fatal(err)
-		}
-		dec.reset()
-		cur = 0 // reset reader
-	}
-}
-
-func BenchmarkDecodeWithFiledef(b *testing.B) {
-	b.StopTimer()
-	// This is not a typical FIT in term of file size (2.3M) and the messages size.
-	// But since it's big, it's should be good to benchmark.
-	f, err := os.Open("../testdata/big_activity.fit")
-	// f, err := os.Open("../testdata/from_official_sdk/activity_lowbattery.fit")
-	if err != nil {
-		panic(err)
-	}
-	defer f.Close()
-
-	all, err := io.ReadAll(f)
-	if err != nil {
-		panic(err)
-	}
-
-	lis := filedef.NewListener()
-	defer lis.Close()
-
-	cur := 0
-	r := fnReader(func(b []byte) (n int, err error) {
-		if cur == len(all) {
-			return 0, io.EOF
-		}
-		n = copy(b, all[cur:])
-		cur += n
-		return
-	})
-
-	dec := New(r,
-		WithMesgListener(lis),
-		WithBroadcastOnly(),
-	)
-	b.StartTimer()
-
-	for i := 0; i < b.N; i++ {
-		_, err = dec.Decode()
-		if err != nil {
-			b.Fatal(err)
-		}
-		_ = lis.File()
-		dec.reset()
-		cur = 0 // reset reader
-	}
-}
-
-func BenchmarkCheckIntegrity(b *testing.B) {
-	b.StopTimer()
-	// This is not a typical FIT in term of file size (2.3M) and the messages size.
-	// But since it's big, it's should be good to benchmark.
-	f, err := os.Open("../testdata/big_activity.fit")
-	if err != nil {
-		panic(err)
-	}
-	defer f.Close()
-
-	all, err := io.ReadAll(f)
-	if err != nil {
-		panic(err)
-	}
-
-	cur := 0
-	r := fnReader(func(b []byte) (n int, err error) {
-		if cur == len(all) {
-			return 0, io.EOF
-		}
-		n = copy(b, all[cur:])
-		cur += n
-		return
-	})
-
-	dec := New(r)
-
-	b.StartTimer()
-
-	for i := 0; i < b.N; i++ {
-		_, err = dec.CheckIntegrity()
-		if err != nil {
-			b.Fatal(err)
-		}
-		cur = 0 // reset reader
-	}
-}
-
-func BenchmarkReset(b *testing.B) {
-	b.Run("benchmark New()", func(b *testing.B) {
-		for i := 0; i < b.N; i++ {
-			_ = New(nil)
-		}
-	})
-	b.Run("benchmark Reset()", func(b *testing.B) {
-		b.StopTimer()
-		dec := New(nil)
-		b.StartTimer()
-
-		for i := 0; i < b.N; i++ {
-			dec.Reset(nil)
-		}
-	})
 }
